debugging output.cpp

+++ src/c/breseq/breseq/output.h
@@ -166,7 +166,7 @@ struct Html_Mutation_Table_String : public string
     //!Helper Functions
     string freq_to_string(const string& freq);//!< Used in Item_Lines()
     string freq_cols(vector<string> freq_list);//!< Used in Item_Lines()
-    uint8_t total_cols; //!< Shared between Factory Methods, set in Header_Line()
+    size_t total_cols; //!< Shared between Factory Methods, set in Header_Line()
 
     //!Parameters
     Settings settings;
